Fixed MRO:ADC_TIMING_SETTING label conversion issue #4290 and converted hi2isis App and Tests
