+++ isis/src/mro/apps/hi2isis/hi2isis.h
@@ -1,0 +1,11 @@
+#ifndef hi2isis_h
+#define hi2isis_h
+
+#include "Cube.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void hi2isis(UserInterface &ui, Pvl *log=nullptr);
+}
+
+#endif
